refactor(voice): split websocket session runtime
